:art: 단일 레포지토리 멀티 프로젝트 구조를 위한 구조 변경
1. 기존 프로젝트 /todo 폴더로 이동
2. 루트폴더에 scripts 폴더 추가 (추후 실행 파일은 모두 여기로)
