--- app/schemas/todo.py
@@ -1,7 +1,0 @@
-from pydantic import BaseModel
-from app.schemas.item import Item
-
-
-class Todo(BaseModel):
-    id: int
-    item: Item
