--- app/schemas/todo_item.py
@@ -1,12 +1,0 @@
-from pydantic import BaseModel
-
-
-class TodoItem(BaseModel):
-    item: str
-
-    class Config:
-        json_schema_extra = {
-            "example": {
-                "item": "..?"
-            }
-        }
