+++ todo/app/schemas/__init__.py
@@ -1,3 +1,3 @@
 from .todo import Todo
 from .item import Item
-from .todo_item import TodoItem
+from .todo_item import TodoItem, TodoItems
