+++ todo/pyproject.toml
@@ -9,6 +9,8 @@ readme = "README.md"
 python = "^3.12"
 fastapi = "^0.111.0"
 uvicorn = "^0.30.1"
+jinja2 = "^3.1.4"
+python-multipart = "^0.0.9"
 
 
 [build-system]
