+++ todo/templates/todo.html
@@ -1,0 +1,47 @@
+{% extends "home.html" %}
+
+{% block todo_container %}
+<main class="container">
+    <hr>
+    <section class="container-fluid">
+        <form method="post", action="./add/">
+            <div class="col-auto">
+                <div class="input-group mb-3">
+                    <input aria-describedby="button-addon2" aria-label="Add a todo" class="form-control" name="item"
+                           placeholder="Purchase Packt's Python workshop course" type="text"
+                           value="{{ item }}"/>
+                    <button class="btn btn-outline-primary" data-mdb-ripple-color="dark" id="button-addon2"
+                            type="submit">
+                        Add Todo
+                    </button>
+                </div>
+            </div>
+        </form>
+    </section>
+    {% if todo %}
+    <article class="card container-fluid">
+        <br/>
+        <h4>Todo ID: {{ todo.id }} </h4>
+        <p>
+            <strong>
+                Item: {{ todo.item }}
+            </strong>
+        </p>
+    </article>
+    {% else %}
+    <section class="container-fluid">
+        <h2 align="center">Todos</h2>
+        <br>
+        <div class="card">
+            <ul class="list-group list-group-flush">
+                {% for todo in todos %}
+                <li class="list-group-item">
+                    {{ loop.index }}. <a href="./{{ loop.index }}"> {{ todo.item }} </a>
+                </li>
+                {% endfor %}
+            </ul>
+        </div>
+        {% endif %}
+    </section>
+</main>
+{% endblock %}
